XtallographyUtils

Documentation for XtallographyUtils.

XtallographyUtils.XtallographyUtils

XtallographyUtils.BaseCentered

XtallographyUtils.BodyCentered

XtallographyUtils.Centering

XtallographyUtils.Cubic

XtallographyUtils.CubicLatticeConstants

XtallographyUtils.FaceCentered

XtallographyUtils.Hexagonal

XtallographyUtils.HexagonalLatticeConstants

XtallographyUtils.LatticeConstants

XtallographyUtils.LatticeConstants

XtallographyUtils.LatticeSystem

XtallographyUtils.Monoclinic

XtallographyUtils.MonoclinicLatticeConstants

XtallographyUtils.Orthorhombic

XtallographyUtils.OrthorhombicLatticeConstants

XtallographyUtils.Primitive

XtallographyUtils.Rhombohedral

XtallographyUtils.RhombohedralLatticeConstants

XtallographyUtils.Tetragonal

XtallographyUtils.TetragonalLatticeConstants

XtallographyUtils.Triclinic

XtallographyUtils.TriclinicLatticeConstants

XtallographyUtils.UnitCell

XtallographyUtils.acos_

XtallographyUtils.asin_

XtallographyUtils.basis

XtallographyUtils.conventional_cell

XtallographyUtils.convert_to_base_centering

XtallographyUtils.convert_to_body_centering

XtallographyUtils.convert_to_mC

XtallographyUtils.convert_to_mI

XtallographyUtils.convert_to_mP

XtallographyUtils.is_basis

XtallographyUtils.is_bravais_lattice

XtallographyUtils.is_equivalent_unit_cell

XtallographyUtils.is_equivalent_unit_cell

XtallographyUtils.is_supercell

XtallographyUtils.is_triclinic_type_I_cell

XtallographyUtils.lattice_system

XtallographyUtils.reduced_cell

XtallographyUtils.satisfies_triclinic_angle_constraints

XtallographyUtils.standardize

XtallographyUtils.standardize

XtallographyUtils.standardize

XtallographyUtils.surface_area

XtallographyUtils.volume

XtallographyUtils.XtallographyUtils — Module

The XtallographyUtils package provides utility functions for analysis of 3D crystals.

XtallographyUtils.BaseCentered — Type

BaseCentered

Type representing base centering

Note

By convention, base-centering is on the C-face of the unit cell.

Supertype: Centering

XtallographyUtils.BodyCentered — Type

BodyCentered

Type representing body centering

Supertype: Centering

XtallographyUtils.Centering — Type

Centering

Supertype for the four lattice centerings in 3D

Subtypes

Primitive, BaseCentered, BodyCentered, FaceCentered

XtallographyUtils.Cubic — Type

Cubic

Type representing the cubic lattice system

Supertype: LatticeSystem

XtallographyUtils.CubicLatticeConstants — Type

CubicLatticeConstants

Lattice constants for a cubic unit cell

Fields

a: length of the edge of the unit cell

Supertype: LatticeConstants

XtallographyUtils.FaceCentered — Type

FaceCentered

Type representing face centering

Supertype: Centering

XtallographyUtils.Hexagonal — Type

Hexagonal

Type representing the hexagonal lattice system

Supertype: LatticeSystem

XtallographyUtils.HexagonalLatticeConstants — Type

HexagonalLatticeConstants

Lattice constants for a hexagonal unit cell

Fields

a, c: lengths of the edges of the unit cell

Supertype: LatticeConstants

XtallographyUtils.LatticeConstants — Type

LatticeConstants

Supertype for lattice constants for the seven lattice systems in 3D

Subtypes

TriclinicLatticeConstants, MonoclinicLatticeConstants, OrthorhombicLatticeConstants, TetragonalLatticeConstants, RhombohedralLatticeConstants, HexagonalLatticeConstants, CubicLatticeConstants

XtallographyUtils.LatticeConstants — Method

LatticeConstants(
    basis_a::Vector{<:Real},
    basis_b::Vector{<:Real},
    basis_c::Vector{<:Real};
    identify_lattice_system=true,
    centering=Primitive()
) -> LatticeConstants

Construct a LatticeConstants object from a set of basis vectors.

Keyword Arguments

identify_lattice_system: if set to true, return a LatticeConstants object for the highest symmetry lattice system that is consistent with the basis. Otherwise, return a TriclinicLatticeConstants object.

centering: centering of unit cell

Examples

TODO

XtallographyUtils.LatticeSystem — Type

LatticeSystem

Supertype for the seven lattice systems in 3D

Subtypes

Triclinic, Monoclinic, Orthorhombic, Hexagonal, Rhombohedral, Tetragonal, Cubic

XtallographyUtils.Monoclinic — Type

Monoclinic

Type representing the monoclinic lattice system

Supertype: LatticeSystem

XtallographyUtils.MonoclinicLatticeConstants — Type

MonoclinicLatticeConstants

Lattice constants for a monoclinic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

β: angle between edges of the unit cell in the plane of the face of the unit cell where the edges are not orthogonal

Supertype: LatticeConstants

XtallographyUtils.Orthorhombic — Type

Orthorhombic

Type representing the orthorhombic lattice system

Supertype: LatticeSystem

XtallographyUtils.OrthorhombicLatticeConstants — Type

OrthorhombicLatticeConstants

Lattice constants for an orthorhombic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

Note

By convention, edge lengths of the unit cell are ordered so that a <= b <= c.

Supertype: LatticeConstants

XtallographyUtils.Primitive — Type

Primitive

Type representing no centering

Supertype: Centering

XtallographyUtils.Rhombohedral — Type

Rhombohedral

Type representing the rhombohedral lattice system

Supertype: LatticeSystem

XtallographyUtils.RhombohedralLatticeConstants — Type

RhombohedralLatticeConstants

Lattice constants for a rhombohedral unit cell

Fields

a: length of the edge of the unit cell

α: angle between edges of the unit cell in the plane of the faces of the unit cell

Supertype: LatticeConstants

XtallographyUtils.Tetragonal — Type

Tetragonal

Type representing the tetragonal lattice system

Supertype: LatticeSystem

XtallographyUtils.TetragonalLatticeConstants — Type

TetragonalLatticeConstants

Lattice constants for a tetragonal unit cell

Fields

a, c: lengths of the edges of the unit cell

Supertype: LatticeConstants

XtallographyUtils.Triclinic — Type

Triclinic

Type representing the triclinic lattice system

Supertype: LatticeSystem

XtallographyUtils.TriclinicLatticeConstants — Type

TriclinicLatticeConstants

Lattice constants for a triclinic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

α, β, γ: angles between edges of the unit cell in the planes of the faces of the unit cell

Note

The constraints that valid triclinic unit cell angles must satisfy are not enforced by the constructor. It is acceptable to construct TriclinicLatticeConstants with invalid values for α, β, and γ.

Supertype: LatticeConstants

XtallographyUtils.UnitCell — Type

UnitCell

Unit cell for a lattice

Fields

lattice_constants: lattice constants of unit cell

centering: centering of unit cell

XtallographyUtils.acos_ — Method

acos_(x::Real; rtol::Real) -> Float64

Compute the arccos of x with a tolerance for values of x that are slightly outside of the mathematical domain [-1, 1].

When the value of x is approximately equal to 1, acos_(x) returns 0; when the value of x is approximately equal to -1, acos_(x) returns π.

Return values

arccos of x

Examples

julia> acos_(0.5) ≈ π / 3
true
julia> acos_(1 + eps(1.0)) == 0
true
julia> acos_(-1 - eps(1.0)) == π
true

XtallographyUtils.asin_ — Method

asin_(x::Real; rtol::Real) -> Float64

Compute the arcsin of x with a tolerance for values of x that are slightly outside of the mathematical domain [-1, 1].

When the value of x is approximately equal to 1, asin_(x) returns π / 2; when the value of x is approximately equal to -1, asin_(x) returns -π / 2.

Return values

arcsin of x

Examples

julia> asin_(0.5) ≈ π / 6
true
julia> asin_(1 + eps(1.0)) == π / 2
true
julia> asin_(-1 - eps(1.0)) == -π / 2
true

XtallographyUtils.basis — Function

basis(unit_cell::UnitCell) -> (Vector{Float64}, Vector{Float64}, Vector{Float64})

basis(
    lattice_constants::LatticeConstants
) -> (Vector{Float64}, Vector{Float64}, Vector{Float64})

Return a set of basis vectors $\vec{a}, \vec{b}, \vec{c}$ for the unit cell defined by unit_cell or lattice_constants.

Return values

basis vectors $\vec{a}$, $\vec{b}$, $\vec{c}$

Examples

TODO

XtallographyUtils.conventional_cell — Method

conventional_cell(unit_cell::UnitCell) -> UnitCell

Return the IUCr conventional cell that is equivalent to unit_cell.

Return values

IUCr conventional cell for unit_cell

Examples

TODO

XtallographyUtils.convert_to_base_centering — Method

convert_to_base_centering(
    lattice_constants::MonoclinicLatticeConstants
) -> MonoclinicLatticeConstants

Convert a body-centered monoclinic unit cell to base-centered monoclinic unit cell.

Return values

lattice constants for equivalent base-centered unit cell

Examples

TODO

XtallographyUtils.convert_to_body_centering — Method

convert_to_body_centering(
    lattice_constants::MonoclinicLatticeConstants
) -> MonoclinicLatticeConstants

Convert a base-centered monoclinic unit cell to body-centered monoclinic unit cell.

Return values

lattice constants for equivalent body-centered unit cell

Examples

TODO

XtallographyUtils.convert_to_mC — Method

convert_to_mC(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent base-centered monoclinic unit cell.

Return values

lattice constants for the equivalent base-centered monoclinic unit cell if one exists; nothing otherwise

Warn

Returned lattice constants are not guaranteed to be standardized.

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a base-centered monoclinic unit cell.

XtallographyUtils.convert_to_mI — Method

convert_to_mI(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent body-centered monoclinic unit cell.

Return values

lattice constants for the equivalent body-centered monoclinic unit cell if one exists; nothing otherwise

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a body-centered monoclinic unit cell.

XtallographyUtils.convert_to_mP — Method

convert_to_mP(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent primitive monoclinic unit cell.

Return values

lattice constants for the equivalent primitive monoclinic unit cell if one exists

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a primitive monoclinic unit cell.

XtallographyUtils.is_basis — Method

is_basis(
    v1::Vector{Real}, v2::Vector{Real}, v3::Vector{Real}
) -> Bool

Determine if the vectors v1, v2, and v3 are a basis for a three-dimensional lattice.

Return values

true if v1, v2, and v3 are a basis; false otherwise

Examples

TODO

XtallographyUtils.is_bravais_lattice — Method

is_bravais_lattice(lattice_system::LatticeSystem, centering::Centering) -> Bool

is_bravais_lattice(lattice_system::Type{<:LatticeSystem}, centering::Centering) -> Bool

is_bravais_lattice(unit_cell::UnitCell) -> Bool

Determine if the unit cell defined by unit_cell or lattice_system and centering is a valid Bravais lattice type.

Return values

true if lattice_system and centering define a valid Bravais lattice type; false otherwise

Examples

TODO

XtallographyUtils.is_equivalent_unit_cell — Method

is_equivalent_unit_cell(
    lattice_constants_test::LatticeConstants,
    lattice_constants_ref::LatticeConstants;
    tol::Real=1e-3
) -> Bool

Check if the primitive unit cell defined by lattice_constants_test is equivalent to the unit cell defined by lattice_constants_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between the reduced unit cells defined by lattice_constants_test and lattice_constants_ref.

Return values

true if the test unit cell is equivalent to the reference unit cell; false otherwise

Examples

TODO

XtallographyUtils.is_equivalent_unit_cell — Method

is_equivalent_unit_cell(
    unit_cell_test::UnitCell,
    unit_cell_ref::UnitCell;
    tol::Real=1e-3
) -> Bool

Check if the unit cell defined by unit_cell_test is equivalent to the unit cell defined by unit_cell_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between the reduced unit cells defined by unit_cell_test and unit_cell_ref.

Return values

true if the test unit cell is equivalent to the reference unit cell; false otherwise

Examples

TODO

XtallographyUtils.is_supercell — Function

is_supercell(
    lattice_constants_test::LatticeConstants,
    lattice_constants_ref::LatticeConstants;
    tol::Real=1e-3,
    max_index::Integer=3
) -> Bool

Check if the unit cell defined by lattice_constants_test is a supercell of the unit cell defined by lattice_constants_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between test unit cell and supercells of the reference unit cell

max_index: maximum multiple of the basis vectors of the reference unit cell to include when checking whether the test unit cell is a supercell of the reference unit cell

Note

max_index is ignored for Cubic lattice systems.

Return values

true if the test unit cell is a supercell of the reference unit cell; false otherwise

Examples

TODO

XtallographyUtils.is_triclinic_type_I_cell — Method

is_triclinic_type_I_cell(lattice_constants::TriclinicLatticeConstants) -> Bool

Determine whether the unit cell defined by lattice_constants is a Type I or Type II cell.

Return values

true if lattice_constants defines a Type I cell; false if lattice_constants defines a Type II cell.

XtallographyUtils.lattice_system — Function

lattice_system(lattice_constants::LatticesConstants) -> LatticeSystem

Return the lattice system for a set of lattice_constants.

Return values

lattice system

XtallographyUtils.reduced_cell — Method

reduced_cell(unit_cell::UnitCell) -> UnitCell

Compute the primitive reduced cell for the lattice defined by unit_cell. The Selling-Delaunay reduction algorithm is used to compute the reduced basis.

Return values

primitive reduced cell

Examples

TODO

XtallographyUtils.satisfies_triclinic_angle_constraints — Method

satisfies_triclinic_angle_constraints(α::Real, β::Real, γ::Real) -> Bool

Determine whether α, β, and γ satisfy the angle constraints for triclinic lattices.

Return values

true if (α, β, γ) form a valid triple of angles for a triclinic unit cell; false otherwise

Examples

julia> satisfies_triclinic_angle_constraints(π/4, π/5, π/6)
true
julia> satisfies_triclinic_angle_constraints(3π/4, 4π/5, 5π/6)
false

XtallographyUtils.standardize — Method

standardize(
    lattice_constants::LatticeConstants, centering::Centering
) -> (LatticeConstants, Centering)

Standardize the lattice constants and centering for the unit cell defined by lattice_constants and centering.

Return values

standardized lattice constants and centering

Examples

TODO

XtallographyUtils.standardize — Method

standardize(lattice_constants::LatticeConstants) -> LatticeConstants

Standardize the lattice constants the primitive unit cell defined by lattice_constants.

Return values

standardized lattice constants for primitive unit cell

Examples

TODO

XtallographyUtils.standardize — Method

standardize(unit_cell::UnitCell) -> LatticeConstants

Standardize the lattice constants and centering for unit_cell.

Note

This function only enforces lattice constant constraints. It does not modify the Bravais lattice type. To find an equivalent Bravais lattice with higher symmetry (if one exists), use conventional_cell().

Note

Lattice constant standardizations are based on the conventions provided in the Table 3.1.4.1. of the International Tables for Crystallography (2016). For triclinic lattices, the lattice constants are standardized using the following conventions:

a ≤ b ≤ c

all three angles are acute (Type I cell) or all three angles are non-acute (Type II cell)

angles sorted in increasing order when edge lengths are equal

α ≤ β when a = b

β ≤ γ when b = c

α ≤ β ≤ γ when a = b = c

Note

Except for monoclinic lattices, standardize() does not modify the unit cell.

For monoclinic lattices, the unit cell may be potentially modified in two ways:

the 2D unit cell in the plane normal to the b-axis may be reduced in order to satisfy the IUCr conventions for a, c, and β;

base-centered unit cells are transformed to equivalent body-centered unit cells.

Return values

unit cell with standardized lattice constants and centering

Examples

TODO

XtallographyUtils.surface_area — Method

surface_area(unit_cell::UnitCell) -> Float64

surface_area(lattice_constants::LatticeConstants) -> Float64

surface_area(
    basis_a::Vector{<:Real}, basis_b::Vector{<:Real}, basis_c::Vector{<:Real}
) -> Float64

Compute the surface area of the unit cell defined by unit_cell, lattice_constants, or the basis vectors [basis_a, basis_b, basis_c].

Return values

surface area of the unit cell

Examples

julia> S = surface_area([1, 0, 0], [1, 1, 0], [1, 0, 1]);

julia> S ≈ 4 * (1 + sqrt(3)/2)
true

XtallographyUtils.volume — Method

volume(unit_cell::UnitCell) -> Float64

volume(lattice_constants::LatticeConstants) -> Float64

volume(
    basis_a::Vector{<:Real}, basis_b::Vector{<:Real}, basis_c::Vector{<:Real}
) -> Float64

Compute the volume of the unit cell defined by unit_cell, lattice_constants, or the basis vectors [basis_a, basis_b, basis_c].

Return values

volume of the unit cell

Examples

TODO