Symmetry

Types and functions related to defining crystal symmetry.

Lattice Systems

Note

All concrete lattice system types are (1) subtypes of the LatticeSystem abstract type and (2) singleton types. For convenience, a singleton instance is defined for each lattice system type.

Xtallography.LatticeSystem — Type

LatticeSystem

Supertype for the seven lattice systems in 3D

Subtypes

Triclinic, Monoclinic, Orthorhombic, Tetragonal, Rhombohedral, Hexagonal, Cubic

Xtallography.lattice_system — Function

lattice_system(unit_cell::UnitCell) -> LatticeSystem

lattice_system(unit_cell::UnitCellDelta) -> LatticeSystem

Return the lattice system for unit_cell or Δunit_cell.

Return values

lattice system

Examples

julia> lattice_system(HexagonalUnitCell(2, 4))
Hexagonal()

julia> lattice_system(CubicUnitCell(2))
Cubic()

julia> lattice_system(OrthorhombicUnitCell(2, 3, 4; centering=F_centering))
Orthorhombic()

julia> lattice_system(TetragonalUnitCellDelta(0.1, .05))
Tetragonal()

Xtallography.LATTICE_SYSTEMS — Constant

LATTICE_SYSTEMS

List of valid lattice systems in 3D

Concrete Types

Xtallography.Cubic — Type

Cubic

Type representing the cubic lattice system that is the type of cubic

Supertype: LatticeSystem

Xtallography.cubic — Constant

cubic

The singleton instance of type Cubic

Xtallography.Hexagonal — Type

Hexagonal

Type representing the hexagonal lattice system that is the type of hexagonal

Supertype: LatticeSystem

Xtallography.hexagonal — Constant

hexagonal

The singleton instance of type Hexagonal

Xtallography.Monoclinic — Type

Monoclinic

Type representing the monoclinic lattice system that is the type of monoclinic

Supertype: LatticeSystem

Xtallography.monoclinic — Constant

monoclinic

The singleton instance of type Monoclinic

Xtallography.Orthorhombic — Type

Orthorhombic

Type representing the orthorhombic lattice system that is the type of orthorhombic

Supertype: LatticeSystem

Xtallography.orthorhombic — Constant

orthorhombic

The singleton instance of type Orthorhombic

Xtallography.Rhombohedral — Type

Rhombohedral

Type representing the rhombohedral lattice system that is the type of rhombohedral

Supertype: LatticeSystem

Xtallography.rhombohedral — Constant

rhombohedral

The singleton instance of type Rhombohedral

Xtallography.Tetragonal — Type

Tetragonal

Type representing the tetragonal lattice system that is the type of tetragonal

Supertype: LatticeSystem

Xtallography.tetragonal — Constant

tetragonal

The singleton instance of type Tetragonal

Xtallography.Triclinic — Type

Triclinic

Type representing the triclinic lattice system that is the type of triclinic

Supertype: LatticeSystem

Xtallography.triclinic — Constant

triclinic

The singleton instance of type Triclinic

Centerings

Note

All concrete centering types are (1) subtypes of the Centering abstract type and (2) singleton types. For convenience, a singleton instance is defined for each centering type.

Xtallography.Centering — Type

Centering

Supertype for the four lattice centerings in 3D

Subtypes

PrimitiveCentering, BaseCentering, BodyCentering, FaceCentering

Xtallography.CENTERINGS — Constant

CENTERINGS

List of valid lattice centerings in 3D

Concrete Types

Xtallography.BaseCentering — Type

BaseCentering

Type representing base centering that is the type of base_centering

Note

By convention, base-centering is

on the C-face of the unit cell for orthorhombic lattice systems

and

on the B-face of the unit cell for monoclinic lattice systems.

Supertype: Centering

Xtallography.base_centering — Constant

base_centering

The singleton instance of type BaseCentering

Xtallography.BodyCentering — Type

BodyCentering

Type representing body centering that is the type of body_centering and I_centering

Supertype: Centering

Xtallography.body_centering — Constant

body_centering

The singleton instance of type BodyCentering

Xtallography.I_centering — Constant

I_centering

Alias for body_centering

Xtallography.FaceCentering — Type

FaceCentering

Type representing face centering that is the type of face_centering and F_centering

Supertype: Centering

Xtallography.face_centering — Constant

face_centering

The singleton instance of type FaceCentering

Xtallography.F_centering — Constant

F_centering

Alias for face_centering

Xtallography.PrimitiveCentering — Type

PrimitiveCentering

Type representing no centering that is the type of primitive_centering and P_centering

Supertype: Centering

Xtallography.primitive_centering — Constant

primitive_centering

The singleton instance of type PrimitiveCentering

Xtallography.P_centering — Constant

P_centering

Alias for primitive_centering

Symmetry Elements

Note

All concrete symmetry element types are subtypes of the SymmetryElement abstract type.

Xtallography.SymmetryElement — Type

SymmetryElement

Supertype for symmetry elements in 3D

Subtypes

RotationAxis, MirrorPlane, InversionCenter, RotoinversionAxis GlidePlane, ScrewAxis

Concrete Types

Xtallography.GlidePlane — Type

GlidePlane

Type representing a glide plane

Supertype: SymmetryElement

Xtallography.GlidePlane — Method

GlidePlane(
    glide::Tuple{<:Real,<:Real,<:Real},
    normal::Tuple{<:Real,<:Real,<:Real};
    location::Tuple{<:Real,<:Real,<:Real}=(0, 0, 0)
)

Construct a GlidePlane object with the specified glide translation and mirror plane with orientation defined by normal.

Keyword Arguments

location: a point on the rotation axis in unit cell fractional coordinates

Xtallography.InversionCenter — Type

InversionCenter

Type representing an inversion through a point

Supertype: SymmetryElement

Xtallography.MirrorPlane — Type

MirrorPlane

Type representing a reflection through a plane

Supertype: SymmetryElement

Xtallography.MirrorPlane — Method

MirrorPlane(
    normal::Tuple{<:Real,<:Real,<:Real};
    location::Tuple{<:Real,<:Real,<:Real}=(0, 0, 0)
)

Construct a MirrorPlane object with orientation defined by normal.

Keyword Arguments

location: a point on the rotation axis in unit cell fractional coordinates

Xtallography.RotationAxis — Type

RotationAxis

Type representing a rotation axis

Supertype: SymmetryElement

Xtallography.RotationAxis — Method

RotationAxis(
    n::Int,
    direction::Tuple{<:Real,<:Real,<:Real};
    location::Tuple{<:Real,<:Real,<:Real}=(0, 0, 0)
)

Construct a RotationAxis object with order n about the axis specified by direction.

Keyword Arguments

location: a point on the rotation axis in unit cell fractional coordinates

Xtallography.RotoinversionAxis — Type

RotoinversionAxis

Type representing a rotoinversion axis

Supertype: SymmetryElement

Xtallography.RotoinversionAxis — Method

RotoinversionAxis(
    n::Int,
    direction::Tuple{<:Real,<:Real,<:Real};
    center::Tuple{<:Real,<:Real,<:Real}=(0, 0, 0)
)

Construct a RotoinversionAxis object with order n about the axis specified by direction.

Keyword Arguments

center: location of inversion center in unit cell fractional coordinates

Xtallography.ScrewAxis — Type

ScrewAxis

Type representing a screw axis

Supertype: SymmetryElement

Xtallography.ScrewAxis — Method

ScrewAxis(
    n::Int,
    m::Int,
    direction::Tuple{<:Real,<:Real,<:Real};
    location::Tuple{<:Real,<:Real,<:Real}=(0, 0, 0)
)

Construct a ScrewAxis object that

has an axis by direction and location

and

couples a rotation by 2π/n and a translation along the rotation axis by m/n (in unit cell fractional coordinates).

Keyword Arguments

location: a point on the rotation axis in unit cell fractional coordinates

Bravais Lattices

Xtallography.BRAVAIS_LATTICES — Constant

BRAVAIS_LATTICES

List of valid Bravais lattices.

Xtallography.is_bravais_lattice — Function

is_bravais_lattice(lattice_system::LatticeSystem, centering::Centering) -> Bool

Determine if lattice_system and centering define a valid Bravais lattice type.

Return values

true if lattice_system and centering define a valid Bravais lattice type; false otherwise

Examples

julia> is_bravais_lattice(cubic, body_centering)
true

julia> is_bravais_lattice(cubic, base_centering)
false

is_bravais_lattice(unit_cell::UnitCell) -> Bool

Determine if the unit cell defined by unit_cell is a valid Bravais lattice type.

Return values

true if the lattice system and centering of unit_cell define a valid Bravais lattice type; false otherwise

Examples

julia> is_bravais_lattice(TetragonalUnitCell(2, 3; centering=primitive_centering))
true

julia> is_bravais_lattice(TetragonalUnitCell(2, 3; centering=face_centering))
false