Utility Functions

Math Functions

XtallographyUtils.asin_ — Function

asin_(x::Real; rtol::Real) -> Float64

Compute the arcsin of x with a tolerance for values of x that are slightly outside of the mathematical domain [-1, 1].

When the value of x is approximately equal to 1, asin_(x) returns π / 2; when the value of x is approximately equal to -1, asin_(x) returns -π / 2.

Return values

arcsin of x

Examples

julia> asin_(0.5) ≈ π / 6
true
julia> asin_(1 + eps(1.0)) == π / 2
true
julia> asin_(-1 - eps(1.0)) == -π / 2
true

XtallographyUtils.acos_ — Function

acos_(x::Real; rtol::Real) -> Float64

Compute the arccos of x with a tolerance for values of x that are slightly outside of the mathematical domain [-1, 1].

When the value of x is approximately equal to 1, acos_(x) returns 0; when the value of x is approximately equal to -1, acos_(x) returns π.

Return values

arccos of x

Examples

julia> acos_(0.5) ≈ π / 3
true
julia> acos_(1 + eps(1.0)) == 0
true
julia> acos_(-1 - eps(1.0)) == π
true

XtallographyUtils.is_basis — Function

is_basis(v1::Vector{Real}, v2::Vector{Real}, v3::Vector{Real}) -> Bool

Determine if the vectors v1, v2, and v3 are a basis for a three-dimensional lattice (i.e., v1, v2, and v3 are linearly independent).

Return values

true if v1, v2, and v3 are a basis; false otherwise

Examples

TODO

XtallographyUtils.volume — Method

volume(v1::Vector{Real}, v2::Vector{Real}, v3::Vector{Real}) -> Float64

Compute the volume of the parallelipiped defined by the vectors v1, v2, and v3.

Return values

volume of the parallelipiped defined by v1, v2, and v3

Examples

TODO

XtallographyUtils.surface_area — Method

surface_area(v1::Vector{Real}, v2::Vector{Real}, v3::Vector{Real}) -> Float64

Compute the surface area of the parallelipiped defined by the vectors v1, v2, and v3.

Return values

surface area of the parallelipiped defined by v1, v2, and v3

Examples

TODO

LatticeConstant Functions

XtallographyUtils.lattice_system — Function

lattice_system(lattice_constants::LatticesConstants) -> LatticeSystem

Return the lattice system for a set of lattice_constants.

Return values

lattice system

XtallographyUtils.standardize — Method

standardize(
    lattice_constants::LatticeConstants, centering::Centering
) -> (LatticeConstants, Centering)

Standardize the lattice constants and centering for the unit cell defined by lattice_constants and centering.

Return values

standardized lattice constants and centering

Examples

TODO

XtallographyUtils.standardize — Method

standardize(lattice_constants::LatticeConstants) -> LatticeConstants

Standardize the lattice constants the primitive unit cell defined by lattice_constants.

Return values

standardized lattice constants for primitive unit cell

Examples

TODO

Unit Cell Function

XtallographyUtils.standardize — Method

standardize(unit_cell::UnitCell) -> LatticeConstants

Standardize the lattice constants and centering for unit_cell.

Note

This function only enforces lattice constant constraints. It does not modify the Bravais lattice type. To find an equivalent Bravais lattice with higher symmetry (if one exists), use conventional_cell().

Note

Lattice constant standardizations are based on the conventions provided in the Table 3.1.4.1. of the International Tables for Crystallography (2016). For triclinic lattices, the lattice constants are standardized using the following conventions:

a ≤ b ≤ c

all three angles are acute (Type I cell) or all three angles are non-acute (Type II cell)

angles sorted in increasing order when edge lengths are equal

α ≤ β when a = b

β ≤ γ when b = c

α ≤ β ≤ γ when a = b = c

Note

Except for monoclinic lattices, standardize() does not modify the unit cell.

For monoclinic lattices, the unit cell may be potentially modified in two ways:

the 2D unit cell in the plane normal to the b-axis may be reduced in order to satisfy the IUCr conventions for a, c, and β;

base-centered unit cells are transformed to equivalent body-centered unit cells.

Return values

unit cell with standardized lattice constants and centering

Examples

TODO

XtallographyUtils.basis — Function

basis(unit_cell::UnitCell) -> (Vector{Float64}, Vector{Float64}, Vector{Float64})

basis(
    lattice_constants::LatticeConstants
) -> (Vector{Float64}, Vector{Float64}, Vector{Float64})

Return a set of basis vectors $\vec{a}, \vec{b}, \vec{c}$ for the unit cell defined by unit_cell or lattice_constants.

Return values

basis vectors $\vec{a}$, $\vec{b}$, $\vec{c}$

Examples

TODO

XtallographyUtils.volume — Method

volume(unit_cell::UnitCell) -> Float64

volume(lattice_constants::LatticeConstants) -> Float64

volume(
    basis_a::Vector{<:Real}, basis_b::Vector{<:Real}, basis_c::Vector{<:Real}
) -> Float64

Compute the volume of the unit cell defined by unit_cell, lattice_constants, or the basis vectors [basis_a, basis_b, basis_c].

Return values

volume of the unit cell

Examples

TODO

XtallographyUtils.surface_area — Method

surface_area(unit_cell::UnitCell) -> Float64

surface_area(lattice_constants::LatticeConstants) -> Float64

surface_area(
    basis_a::Vector{<:Real}, basis_b::Vector{<:Real}, basis_c::Vector{<:Real}
) -> Float64

Compute the surface area of the unit cell defined by unit_cell, lattice_constants, or the basis vectors [basis_a, basis_b, basis_c].

Return values

surface area of the unit cell

Examples

julia> S = surface_area([1, 0, 0], [1, 1, 0], [1, 0, 1]);

julia> S ≈ 4 * (1 + sqrt(3)/2)
true

XtallographyUtils.conventional_cell — Function

conventional_cell(unit_cell::UnitCell) -> UnitCell

Return the IUCr conventional cell that is equivalent to unit_cell.

Return values

IUCr conventional cell for unit_cell

Examples

TODO

XtallographyUtils.reduced_cell — Function

reduced_cell(unit_cell::UnitCell) -> UnitCell

Compute the primitive reduced cell for the lattice defined by unit_cell. The Selling-Delaunay reduction algorithm is used to compute the reduced basis.

Return values

primitive reduced cell

Examples

TODO

XtallographyUtils.is_equivalent_unit_cell — Function

is_equivalent_unit_cell(
    unit_cell_test::UnitCell,
    unit_cell_ref::UnitCell;
    tol::Real=1e-3
) -> Bool

Check if the unit cell defined by unit_cell_test is equivalent to the unit cell defined by unit_cell_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between the reduced unit cells defined by unit_cell_test and unit_cell_ref.

Return values

true if the test unit cell is equivalent to the reference unit cell; false otherwise

Examples

TODO

is_equivalent_unit_cell(
    lattice_constants_test::LatticeConstants,
    lattice_constants_ref::LatticeConstants;
    tol::Real=1e-3
) -> Bool

Check if the primitive unit cell defined by lattice_constants_test is equivalent to the unit cell defined by lattice_constants_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between the reduced unit cells defined by lattice_constants_test and lattice_constants_ref.

Return values

true if the test unit cell is equivalent to the reference unit cell; false otherwise

Examples

TODO

XtallographyUtils.is_supercell — Function

is_supercell(
    lattice_constants_test::LatticeConstants,
    lattice_constants_ref::LatticeConstants;
    tol::Real=1e-3,
    max_index::Integer=3
) -> Bool

Check if the unit cell defined by lattice_constants_test is a supercell of the unit cell defined by lattice_constants_ref.

Keyword Arguments

tol: absolute tolerance of the deviation between test unit cell and supercells of the reference unit cell

max_index: maximum multiple of the basis vectors of the reference unit cell to include when checking whether the test unit cell is a supercell of the reference unit cell

Note

max_index is ignored for Cubic lattice systems.

Return values

true if the test unit cell is a supercell of the reference unit cell; false otherwise

Examples

TODO

Triclinic Unit Cell Functions

XtallographyUtils.is_triclinic_type_I_cell — Function

is_triclinic_type_I_cell(lattice_constants::TriclinicLatticeConstants) -> Bool

Determine whether the unit cell defined by lattice_constants is a Type I or Type II cell.

Return values

true if lattice_constants defines a Type I cell; false if lattice_constants defines a Type II cell.

XtallographyUtils.convert_to_body_centering — Function

convert_to_body_centering(
    lattice_constants::MonoclinicLatticeConstants
) -> MonoclinicLatticeConstants

Convert a base-centered monoclinic unit cell to body-centered monoclinic unit cell.

Return values

lattice constants for equivalent body-centered unit cell

Examples

TODO

XtallographyUtils.convert_to_base_centering — Function

convert_to_base_centering(
    lattice_constants::MonoclinicLatticeConstants
) -> MonoclinicLatticeConstants

Convert a body-centered monoclinic unit cell to base-centered monoclinic unit cell.

Return values

lattice constants for equivalent base-centered unit cell

Examples

TODO

XtallographyUtils.convert_to_mP — Function

convert_to_mP(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent primitive monoclinic unit cell.

Return values

lattice constants for the equivalent primitive monoclinic unit cell if one exists

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a primitive monoclinic unit cell.

XtallographyUtils.convert_to_mI — Function

convert_to_mI(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent body-centered monoclinic unit cell.

Return values

lattice constants for the equivalent body-centered monoclinic unit cell if one exists; nothing otherwise

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a body-centered monoclinic unit cell.

XtallographyUtils.convert_to_mC — Function

convert_to_mC(
    lattice_constants::TriclinicLatticeConstants
) -> MonoclinicLatticeConstants

Attempt to convert the triclinic unit cell defined by lattice_constants to an equivalent base-centered monoclinic unit cell.

Return values

lattice constants for the equivalent base-centered monoclinic unit cell if one exists; nothing otherwise

Warn

Returned lattice constants are not guaranteed to be standardized.

Exceptions

Throws an ErrorException if the triclinic unit cell defined by lattice_constants is not equivalent to a base-centered monoclinic unit cell.

XtallographyUtils.satisfies_triclinic_angle_constraints — Function

satisfies_triclinic_angle_constraints(α::Real, β::Real, γ::Real) -> Bool

Determine whether α, β, and γ satisfy the angle constraints for triclinic lattices.

Return values

true if (α, β, γ) form a valid triple of angles for a triclinic unit cell; false otherwise

Examples

julia> satisfies_triclinic_angle_constraints(π/4, π/5, π/6)
true
julia> satisfies_triclinic_angle_constraints(3π/4, 4π/5, 5π/6)
false