Types

Lattice Types

LatticeSystem

Centering

Unit Cell Types

LatticeConstants

LatticeConstantDeltas

UnitCell

Lattice Types

XtallographyUtils.LatticeSystem — Type

LatticeSystem

Supertype for the seven lattice systems in 3D

Subtypes

Triclinic, Monoclinic, Orthorhombic, Hexagonal, Rhombohedral, Tetragonal, Cubic

XtallographyUtils.Centering — Type

Centering

Supertype for the four lattice centerings in 3D

Subtypes

Primitive, BaseCentered, BodyCentered, FaceCentered

Concrete Types

Lattice Systems

All subtypes of LatticeSystem are singleton types. For convenience, a singleton instance is defined for each lattice system type.

XtallographyUtils.Triclinic — Type

Triclinic

Type representing the triclinic lattice system that is the type of triclinic

Supertype: LatticeSystem

XtallographyUtils.Monoclinic — Type

Monoclinic

Type representing the monoclinic lattice system that is the type of monoclinic

Supertype: LatticeSystem

XtallographyUtils.Orthorhombic — Type

Orthorhombic

Type representing the orthorhombic lattice system that is the type of orthorhombic

Supertype: LatticeSystem

XtallographyUtils.Hexagonal — Type

Hexagonal

Type representing the hexagonal lattice system that is the type of hexagonal

Supertype: LatticeSystem

XtallographyUtils.Rhombohedral — Type

Rhombohedral

Type representing the rhombohedral lattice system that is the type of rhombohedral

Supertype: LatticeSystem

XtallographyUtils.Tetragonal — Type

Tetragonal

Type representing the tetragonal lattice system that is the type of tetragonal

Supertype: LatticeSystem

XtallographyUtils.Cubic — Type

Cubic

Type representing the cubic lattice system that is the type of cubic

Supertype: LatticeSystem

Centerings

All subtypes of Centering are singleton types. For convenience, a singleton instance is defined for each centering type.

XtallographyUtils.Primitive — Type

Primitive

Type representing no centering that is the type of primitive

Supertype: Centering

XtallographyUtils.BaseCentered — Type

BaseCentered

Type representing base centering that is the type of base_centered

Note

By convention, base-centering is

on the C-face of the unit cell for orthorhombic lattice systems

and

on the B-face of the unit cell for monoclinic lattice systems.

Supertype: Centering

XtallographyUtils.BodyCentered — Type

BodyCentered

Type representing body centering that is the type of body_centered

Supertype: Centering

XtallographyUtils.FaceCentered — Type

FaceCentered

Type representing face centering that is the type of face_centered

Supertype: Centering

Constants

Lattice Systems

XtallographyUtils.triclinic — Constant

triclinic

The singleton instance of type Triclinic

XtallographyUtils.monoclinic — Constant

monoclinic

The singleton instance of type Monoclinic

XtallographyUtils.orthorhombic — Constant

orthorhombic

The singleton instance of type Orthorhombic

XtallographyUtils.hexagonal — Constant

hexagonal

The singleton instance of type Hexagonal

XtallographyUtils.rhombohedral — Constant

rhombohedral

The singleton instance of type Rhombohedral

XtallographyUtils.tetragonal — Constant

tetragonal

The singleton instance of type Tetragonal

XtallographyUtils.cubic — Constant

cubic

The singleton instance of type Cubic

Centerings

XtallographyUtils.primitive — Constant

primitive

The singleton instance of type Primitive

XtallographyUtils.base_centered — Constant

base_centered

The singleton instance of type BaseCentered

XtallographyUtils.body_centered — Constant

body_centered

The singleton instance of type BodyCentered

XtallographyUtils.face_centered — Constant

face_centered

The singleton instance of type FaceCentered

Other Constants

XtallographyUtils.BRAVAIS_LATTICES — Constant

BRAVAIS_LATTICES

List of valid Bravais lattices.

Unit Cell Types

XtallographyUtils.LatticeConstants — Type

LatticeConstants{T<:LatticeSystem}

Supertype for lattice constants for the seven lattice systems in 3D

Subtypes

TriclinicLatticeConstants, MonoclinicLatticeConstants, OrthorhombicLatticeConstants, TetragonalLatticeConstants, RhombohedralLatticeConstants, HexagonalLatticeConstants, CubicLatticeConstants

XtallographyUtils.LatticeConstantDeltas — Type

LatticeConstantDeltas{T<:LatticeSystem}

Supertype for lattice constant deltas for the seven lattice systems in 3D

Subtypes

TriclinicLatticeConstantDeltas, MonoclinicLatticeConstantDeltas, OrthorhombicLatticeConstantDeltas, TetragonalLatticeConstantDeltas, RhombohedralLatticeConstantDeltas, HexagonalLatticeConstantDeltas, CubicLatticeConstantDeltas

XtallographyUtils.UnitCell — Type

UnitCell

Unit cell for a lattice

Fields

lattice_constants: lattice constants of unit cell

centering: centering of unit cell

Lattice Constants

XtallographyUtils.TriclinicLatticeConstants — Type

TriclinicLatticeConstants

Lattice constants for a triclinic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

α, β, γ: angles between edges of the unit cell in the planes of the faces of the unit cell

Note

The constraints that valid triclinic unit cell angles must satisfy are not enforced by the constructor. It is acceptable to construct TriclinicLatticeConstants with invalid values for α, β, and γ.

Supertype: LatticeConstants

XtallographyUtils.MonoclinicLatticeConstants — Type

MonoclinicLatticeConstants

Lattice constants for a monoclinic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

β: angle between edges of the unit cell in the plane of the face of the unit cell where the edges are not orthogonal

Supertype: LatticeConstants

XtallographyUtils.OrthorhombicLatticeConstants — Type

OrthorhombicLatticeConstants

Lattice constants for an orthorhombic unit cell

Fields

a, b, c: lengths of the edges of the unit cell

Supertype: LatticeConstants

XtallographyUtils.HexagonalLatticeConstants — Type

HexagonalLatticeConstants

Lattice constants for a hexagonal unit cell

Fields

a, c: lengths of the edges of the unit cell

Supertype: LatticeConstants

XtallographyUtils.RhombohedralLatticeConstants — Type

RhombohedralLatticeConstants

Lattice constants for a rhombohedral unit cell

Fields

a: length of the edge of the unit cell

α: angle between edges of the unit cell in the plane of the faces of the unit cell

Supertype: LatticeConstants

XtallographyUtils.TetragonalLatticeConstants — Type

TetragonalLatticeConstants

Lattice constants for a tetragonal unit cell

Fields

a, c: lengths of the edges of the unit cell

Supertype: LatticeConstants

XtallographyUtils.CubicLatticeConstants — Type

CubicLatticeConstants

Lattice constants for a cubic unit cell

Fields

a: length of the edge of the unit cell

Supertype: LatticeConstants

Lattice Constant Deltas

XtallographyUtils.TriclinicLatticeConstantDeltas — Type

TriclinicLatticeConstantDeltas

Lattice constant deltas for a triclinic unit cell

Fields

Δa, Δb, Δc: deltas of the lengths of the edges of the unit cell

Δα, Δβ, Δγ: deltas of the angles between edges of the unit cell in the planes of the faces of the unit cell

Supertype: LatticeConstantDeltas

XtallographyUtils.MonoclinicLatticeConstantDeltas — Type

MonoclinicLatticeConstantDeltas

Lattice constant deltas for a monoclinic unit cell

Fields

Δa, Δb, Δc: deltas of the lengths of the edges of the unit cell

Δβ: delta of the angle between edges of the unit cell in the plane of the face of the unit cell where the edges are not orthogonal

Supertype: LatticeConstantDeltas

XtallographyUtils.OrthorhombicLatticeConstantDeltas — Type

OrthorhombicLatticeConstantDeltas

Lattice constant deltas for a orthorhombic unit cell

Fields

Δa, Δb, Δc: deltas of the lengths of the edges of the unit cell

Supertype: LatticeConstantDeltas

XtallographyUtils.HexagonalLatticeConstantDeltas — Type

HexagonalLatticeConstantDeltas

Lattice constant deltas for a hexagonal unit cell

Fields

Δa, Δc: deltas of the lengths of the edges of the unit cell

Supertype: LatticeConstantDeltas

XtallographyUtils.RhombohedralLatticeConstantDeltas — Type

RhombohedralLatticeConstantDeltas

Lattice constant deltas for a rhombohedral unit cell

Fields

Δa: delta of the length of the edge of the unit cell

Δα: delta of angle between edges of the unit cell in the plane of the faces of the unit cell

Supertype: LatticeConstantDeltas

XtallographyUtils.TetragonalLatticeConstantDeltas — Type

TetragonalLatticeConstantDeltas

Lattice constant deltas for a tetragonal unit cell

Fields

Δa, Δc: deltas of the lengths of the edges of the unit cell

Supertype: LatticeConstantDeltas

XtallographyUtils.CubicLatticeConstantDeltas — Type

CubicLatticeConstantDeltas

Lattice constant deltas for a cubic unit cell

Fields

Δa: delta of the length of the edge of the unit cell

Supertype: LatticeConstantDeltas